requirements.txt
flask==1.1.2
GetOldTweets3==0.0.11
gunicorn==20.0.4
instagram-explore==0.3.0
numpy==1.18.1
opencv-contrib-python==4.1.2.30
Pillow==7.0.0
requests==2.22.0
